+++ spring/src/test/java/cucumber/runtime/java/spring/beans/BellyBean.java
@@ -1,9 +1,9 @@
-package cucumber.runtime.java.spring;
+package cucumber.runtime.java.spring.beans;
 
 
 public class BellyBean {
 
-    private int cukes;
+    private int cukes = 0;
 
     public void setCukes(int cukes) {
         this.cukes = cukes;