+++ spring/src/test/java/cucumber/runtime/java/spring/commonglue/AutowiresThirdStepDef.java
@@ -1,0 +1,14 @@
+package cucumber.runtime.java.spring.commonglue;
+
+import org.springframework.beans.factory.annotation.Autowired;
+
+public class AutowiresThirdStepDef {
+
+    @Autowired
+    private ThirdStepDef thirdStepDef;
+
+    public ThirdStepDef getThirdStepDef() {
+        return thirdStepDef;
+    }
+
+}