+++ spring/src/test/java/cucumber/runtime/java/spring/contextconfig/BellyStepdefs.java
@@ -1,7 +1,9 @@
-package cucumber.runtime.java.spring;
+package cucumber.runtime.java.spring.contextconfig;
 
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
+import cucumber.runtime.java.spring.beans.Belly;
+import cucumber.runtime.java.spring.beans.BellyBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 
@@ -19,6 +21,7 @@ public class BellyStepdefs {
     @Then("^there are (\\d+) cukes in my belly")
     public void checkCukes(final int n) {
         assertEquals(n, belly.getCukes());
+        belly.setCukes(0);
     }
 
     @Given("^I have (\\d+) cukes in my belly")