+++ History.md
@@ -2,6 +2,7 @@
 
 * [Android,Spring,Needle,Examples] Remove commons-logging & log4j and redirect all logging to slf4j & logback ([#742](https://github.com/cucumber/cucumber-jvm/pull/742) Nayan Hajratwala)
 * [Groovy] Support more then one `World {}` definition ([#716](https://github.com/cucumber/cucumber-jvm/pull/716) Anton)
+* [Spring] Fix the glue class autowiring, transaction and cucumber-glue scope issues ([#711](https://github.com/cucumber/cucumber-jvm/pull/711) Björn Rasmusson)
 
 ## [1.1.8](https://github.com/cucumber/cucumber-jvm/compare/v1.1.7...v1.1.8) (2014-06-26)